refactor(documentForm): move context in component
